Allow acks longer than 65535us and specify ack length in the <C ACK MAX 20 MS> format

+++ DCCACK.h
@@ -79,10 +79,10 @@ class DCCACK {
     static inline void setAckLimit(int mA) {
 	ackLimitmA = mA;
     }
-    static inline void setMinAckPulseDuration(unsigned int i) {
+    static inline void setMinAckPulseDuration(unsigned long i) {
 	minAckPulseDuration = i;
     }
-    static inline void setMaxAckPulseDuration(unsigned int i) {
+    static inline void setMaxAckPulseDuration(unsigned long i) {
 	maxAckPulseDuration = i;
     }
 
@@ -126,11 +126,11 @@ class DCCACK {
     static unsigned long ackCheckStart; // millis
     static unsigned int ackCheckDuration; // millis       
     
-    static unsigned int ackPulseDuration;  // micros
+    static unsigned long ackPulseDuration;  // micros
     static unsigned long ackPulseStart; // micros
 
-    static unsigned int minAckPulseDuration ; // micros
-    static unsigned int maxAckPulseDuration ; // micros
+    static unsigned long minAckPulseDuration ; // micros
+    static unsigned long maxAckPulseDuration ; // micros
     static MotorDriver* progDriver;
     static volatile uint8_t numAckGaps;
     static volatile uint8_t numAckSamples;
